// model/client.ts
export interface Client {
    clientId?: number;
    clientName: string;
    clientEmail: string;
    clientPhone: number;
    clientPassword:string

    // constructor(clientId: number, clientName: string, clientEmail: string, clientPhone: number
    // ,clientPassword:string) {
    //     this.clientId = clientId;
    //     this.clientName = clientName;
    //     this.clientEmail = clientEmail;
    //     this.clientPhone = clientPhone;
    //     this.clientPassword = clientPassword;
    // }

}